Create Travel Alert (Issue #254)
The Travel Alert Reminder is a web application that allows users to set alerts for upcoming travels. Users can input a travel alert message and a specific date and time for the reminder. The application maintains a list of alerts, displaying them in order of their scheduled times. When it's time for an alert, a notification pop-up appears to remind the user of their travel plans.